+++ BB/lib/__init__.py
@@ -1,2 +1,2 @@
 # TODO: For some reason, this module doesnt get recognised so i have to manually import it
-from . import jsonHandler
+from . import jsonHandler, timeUtil
